Fixing TZ bug in read output
We were using from utctimestamp, but passing in a tz aware
object.  This was causing the time to be printed in utc, but
still appending the specified tz offset.

We're now passing the tz aware datetime to fromtimestamp.  This
results in the specified tz being printed along with the proper
tz offset.