Corrected issue with reading multiple RIDs
When you try to read from multiple RID's that have data sampled at
different rates, the readmult() function will end up using the timestamp
of the slowest as the starting point for the next chunk of the read.

For example, consider two dataports, the first stores data every 10
second, the second stores data every 60 seconds.  If you try to read
like:

  exo read <cik> --limit=1000 10_sec 60_sec

The readmult() function asks for chunksize datapoints from each and then
combines then together.  Assuming a chunksize of 212, the 10 second
dataport will return data covering 35:20 minutes.  The 60 second
dataport will return data covering 212 minutes.  When those lists are
combined, the last timestamp is used as the starting value for the next
chunk read.  So we'd miss reading the 10_sec data from from 35:21 to
212:00.

To fix this, we:

  - Read each RID up to the requested limit into a list, keeping track
    of the last timestamp seen for each RID as well as the amount of
    data points returned.
  - Combine all RID lists
  - Truncate the combined list to the requested limit